// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  realiseDates: string[];
  rating: number[];
  previewsVideo: string[];
  videoLinks: string[];
  actors: Array<string[]>;
  directors: string[];
  durations: number[];
  commentsCount?: number;
  userNames: string[];
  userAvatars: string[];
  userEmails: string[];
  userPasswords: string[];
  posters: string[];
  backgroundImages: string[];
  backgroundColors: string[];
}